refactor: extract form validation helpers
